+++ apm-web/src/main/java/com/enniu/qa/ptm/model/ApiTestConfig.java
@@ -171,7 +171,7 @@ public class ApiTestConfig extends BaseModel<ApiTestConfig> {
 	private Boolean useFixedRateRPS;
 
 	@Column(name = "rps")
-	private Integer rps;
+	private Double rps;
 
 	@PrePersist
 	@PreUpdate
@@ -206,7 +206,7 @@ public class ApiTestConfig extends BaseModel<ApiTestConfig> {
 		this.rampUpIncrementInterval = getSafe(this.rampUpIncrementInterval, 1000);
 		this.rampUpType = getSafe(this.rampUpType, RampUp.PROCESS);
 		this.useFixedRateRPS=getSafe(this.useFixedRateRPS,false);
-		this.rps=getSafe(this.rps,0);
+		this.rps=getSafe(this.rps,0.0);
 		this.agentId=getSafe(this.agentId,"");
 	}
 
@@ -226,11 +226,11 @@ public class ApiTestConfig extends BaseModel<ApiTestConfig> {
 		this.agentId = agentId;
 	}
 
-	public Integer getRps() {
+	public Double getRps() {
 		return rps;
 	}
 
-	public void setRps(Integer rps) {
+	public void setRps(Double rps) {
 		this.rps = rps;
 	}
